+++ src/Memcached.php
@@ -8,6 +8,12 @@
 
 namespace Joomla\Cache;
 
+use Joomla\Cache\Exception\UnsupportedFormatException;
+use Joomla\Cache\Item\HasExpirationDateInterface;
+use Psr\Cache\CacheItemInterface;
+use Joomla\Cache\Exception\RuntimeException;
+use Joomla\Cache\Item\Item;
+
 /**
  * Memcached cache driver for the Joomla Framework.
  *
@@ -33,7 +39,7 @@ class Memcached extends Cache
 	{
 		if (!extension_loaded('memcached') || !class_exists('Memcached'))
 		{
-			throw new \RuntimeException('Memcached not supported.');
+			throw new UnsupportedFormatException('Memcached not supported.');
 		}
 
 		// Parent sets up the caching options and checks their type
@@ -79,7 +85,7 @@ class Memcached extends Cache
 	 *
 	 * @since   1.0
 	 */
-	public function get($key)
+	public function getItem($key)
 	{
 		$this->connect();
 		$value = $this->driver->get($key);
@@ -88,7 +94,7 @@ class Memcached extends Cache
 
 		if ($code === \Memcached::RES_SUCCESS)
 		{
-			$item->setValue($value);
+			$item->set($value);
 		}
 
 		return $item;
@@ -103,39 +109,46 @@ class Memcached extends Cache
 	 *
 	 * @since   1.0
 	 */
-	public function remove($key)
+	public function deleteItem($key)
 	{
 		$this->connect();
 
-		$this->driver->delete($key);
-
-		$rc = $this->driver->getResultCode();
-
-		if ( ($rc != \Memcached::RES_SUCCESS))
+		if ($this->hasItem($key))
 		{
-// 			throw new \RuntimeException(sprintf('Unable to remove cache entry for %s. Error message `%s`.', $key, $this->driver->getResultMessage()));
-			return false;
+			$this->driver->delete($key);
+
+			$rc = $this->driver->getResultCode();
+
+			if ( ($rc != \Memcached::RES_SUCCESS))
+			{
+				throw new RuntimeException(sprintf('Unable to remove cache entry for %s. Error message `%s`.', $key, $this->driver->getResultMessage()));
+			}
 		}
 
 		return true;
 	}
 
 	/**
-	 * Method to set a value for a storage entry.
+	 * Persists a cache item immediately.
 	 *
-	 * @param   string   $key    The storage entry identifier.
-	 * @param   mixed    $value  The data to be stored.
-	 * @param   integer  $ttl    The number of seconds before the stored data expires.
+	 * @param   CacheItemInterface  $item  The cache item to save.
 	 *
-	 * @return  boolean
-	 *
-	 * @since   1.0
+	 * @return  static  The invoked object.
 	 */
-	public function set($key, $value, $ttl = null)
+	public function save(CacheItemInterface $item)
 	{
 		$this->connect();
 
-		$this->driver->set($key, $value, $ttl);
+		if ($item instanceof HasExpirationDateInterface)
+		{
+			$ttl = $this->convertItemExpiryToSeconds($item);
+		}
+		else
+		{
+			$ttl = 0;
+		}
+
+		$this->driver->set($item->getKey(), $item->get(), $ttl);
 
 		return (bool) ($this->driver->getResultCode() == \Memcached::RES_SUCCESS);
 	}
@@ -149,7 +162,7 @@ class Memcached extends Cache
 	 *
 	 * @since   1.0
 	 */
-	protected function exists($key)
+	public function hasItem($key)
 	{
 		$this->connect();
